+++ src/model/index.ts
@@ -1,5 +1,4 @@
 export * from './Annotation';
-export * from './DataModel';
 export * from './EntityType';
 export * from './Folder';
 export * from './Image';
