+++ src/store/index.ts
@@ -1,2 +1,3 @@
+export * from './DataModel';
 export * from './StoreProvider';
 export type { Store } from './Store';
